requirements.txt
pandas
msoffcrypto-tool
xlrd
pyyaml